feat: Include series data when fetching users in user route
This commit modifies the user route to include the series data when fetching users from the database. By adding the `include` option with `series: true` in the `prisma.user.findMany` method, the series data will be included in the response. This enhancement improves the functionality of the user route by providing additional information about the users' associated series.