feat: Add default value for seasons in Serie model
This commit adds a default value of 1 for the `seasons` field in the `Serie` model. By setting a default value, it ensures that the `seasons` field will always have a value, even if it is not explicitly provided. This enhancement improves the consistency and reliability of the data stored in the `Serie` model.