feat: Add list_series function to retrieve and display series information
This commit adds a new function called `list_series` to the code. This function makes a GET request to the `/series` endpoint and retrieves a list of series data. It then displays the series information, including the ID, name, number of seasons, genre, and streaming platform. This feature enhances the user interface by providing a way to view and browse the available series.